pyproject.toml
[project]
name = "zimmauth"
dynamic = ["version", "description"]
authors = [{name = "Endre Márk Borza", email = "endremborza@gmail.com"}]
license = {file = "LICENSE"}
readme = "README.md"
requires-python = ">=3.8"
dependencies = ["cryptography"]

[project.optional-dependencies]
env = ["toml"]
ssh = ["fabric", "decorator"]
test = [
    "branthebuilder", 
    "mock-ssh-server", 
    "dvc[s3,ssh]>=3.1.0", 
    "moto", 
    "aiobotocore", 
    "boto3", 
    "fabric", 
    "toml", 
    "decorator"
]
doc = ["branthebuilder[doc]"]

[project.urls]
Homepage = "https://github.com/endremborza/zimmauth"

[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[tool.branb]
line-length = 88
